+++ tianshou/policy/modelfree/bcq.py
@@ -1,6 +1,6 @@
 import torch
-import numpy as np
 import torch.nn.functional as F
+import numpy as np
 from typing import Any, Dict, Union, Optional
 
 from tianshou.data import Batch, to_torch
@@ -15,30 +15,6 @@ class BCQPolicy(BasePolicy):
 
     """
 
-    class BCQN(nn.Module):
-        """BCQ NN for dialogue policy. It includes a net for imitation and a net for Q-value"""
-
-        def __init__(
-            self, input_size, n_actions, imitation_model_hidden_dim, policy_model_hidden_dim
-        ):
-            super(BCQPolicy.BCQN, self).__init__()
-            self.q1 = nn.Linear(input_size, policy_model_hidden_dim)
-            self.q2 = nn.Linear(policy_model_hidden_dim, policy_model_hidden_dim)
-            self.q3 = nn.Linear(policy_model_hidden_dim, n_actions)
-
-            self.i1 = nn.Linear(input_size, imitation_model_hidden_dim)
-            self.i2 = nn.Linear(imitation_model_hidden_dim, imitation_model_hidden_dim)
-            self.i3 = nn.Linear(imitation_model_hidden_dim, n_actions)
-
-        def forward(self, state):
-            q = F.relu(self.q1(state))
-            q = F.relu(self.q2(q))
-
-            i = F.relu(self.i1(state))
-            i = F.relu(self.i2(i))
-            i = F.relu(self.i3(i))
-            return self.q3(q), F.log_softmax(i, dim=1), i
-
     def __init__(
         self,
         model: torch.nn.Module,
@@ -52,6 +28,7 @@ class BCQPolicy(BasePolicy):
     ) -> None:
         super().__init__(**kwargs)
         self._policy_net = model
+        self._policy_net.to(device)
         self._optimizer = optim
         self._cnt = 0
         self._device = device
@@ -59,6 +36,7 @@ class BCQPolicy(BasePolicy):
         self._tau = tau
         self._target_net = deepcopy(self._policy_net)
         self._target_net.eval()
+        self._target_net.to(device)
         self._target_update_frequency = target_update_frequency
         self._imitation_logits_penalty = imitation_logits_penalty
 
@@ -70,7 +48,7 @@ class BCQPolicy(BasePolicy):
     ) -> Batch:
         batch.to_torch()
 
-        state = batch.obs
+        state = batch.obs.to(self._device)
         q, imt, _ = self._policy_net(state.float())
         imt = imt.exp()
         imt = (imt / imt.max(1, keepdim=True)[0] > self._tau).float()
@@ -93,12 +71,13 @@ class BCQPolicy(BasePolicy):
             non_final_next_states = torch.cat(
                 [s.obs.unsqueeze(0) for s in batch if not s.done], dim=0
             )
+            non_final_next_states = non_final_next_states.to(self._device)
         except Exception:
             non_final_next_states = None
 
         # Compute the target Q value
         with torch.no_grad():
-            expected_state_action_values = batch.rew.float()
+            expected_state_action_values = batch.rew.float().to(self._device)
 
             # Add target Q value for non-final next_state
             if non_final_next_states is not None:
@@ -108,7 +87,6 @@ class BCQPolicy(BasePolicy):
 
                 # Use large negative number to mask actions from argmax
                 next_action = (imt * q + (1 - imt) * -1e8).argmax(1, keepdim=True)
-
                 q, _, _ = self._target_net(non_final_next_states)
                 q = q.gather(1, next_action).reshape(-1, 1)
 
@@ -118,12 +96,12 @@ class BCQPolicy(BasePolicy):
                 expected_state_action_values += next_state_values * self._gamma
 
         # Get current Q estimate
-        current_Q, imt, i = self._policy_net(batch.obs)
-        current_Q = current_Q.gather(1, batch.act.unsqueeze(1)).squeeze()
+        current_Q, imt, i = self._policy_net(batch.obs.to(self._device))
+        current_Q = current_Q.gather(1, batch.act.unsqueeze(1).to(self._device)).squeeze()
 
         # Compute Q loss
         q_loss = F.smooth_l1_loss(current_Q, expected_state_action_values)
-        i_loss = F.nll_loss(imt, batch.act.reshape(-1))
+        i_loss = F.nll_loss(imt, batch.act.reshape(-1).to(self._device))
 
         Q_loss = q_loss + i_loss + self._imitation_logits_penalty * i.pow(2).mean()
